+++ src/compiler/emitter/mod.rs
@@ -16,5 +16,7 @@ mod lines_4000_4500;
 pub use lines_4000_4500::*;
 mod lines_4500_5000;
 pub use lines_4500_5000::*;
-mod lines_5000_end;
-pub use lines_5000_end::*;
+mod lines_5000_5500;
+pub use lines_5000_5500::*;
+mod lines_5500_end;
+pub use lines_5500_end::*;
